Use hierarchy in channel path (now uses data-array/<sensor>) for data
- Refactors all coral enviro senors into one module
